+++ ceammc/ext/src/noise/noise_lfreq0.cpp
@@ -1,3 +1,11 @@
 #include "noise_lfreq0.h"
 
-SIMPLE_EXTERNAL(noise);
+EXTERNAL_NEW
+{
+    FAUST_EXT* x = reinterpret_cast<FAUST_EXT*>(pd_new(FAUST_EXT_CLASS));
+    PdArgParser p(x, argc, argv);
+    p.initFloatArg("freq", 1);
+    return p.pd_obj();
+}
+
+EXTERNAL_SETUP(noise);