+++ ceammc/CHANGELOG.md
@@ -5,6 +5,8 @@
 - env.asr~ envelope generator added
 - env.smooth~ envelope generator added
 - fx.drone_box~ added
+- noise.lfreq0~ added
+- flt.eq_peak_cq~ added
 
 ### Changed:
 - migration to Faust v2